Use strategy pattern for minmax searches in PlottableSignal classes.
