+++ src/ScottPlot/MinMaxSearchStrategies/SegmentedTreeMinMaxSearchStrategy.cs
@@ -1,0 +1,46 @@
+﻿using System;
+using ScottPlot.DataStructures;
+
+namespace ScottPlot.MinMaxSearchStrategies
+{
+    public class SegmentedTreeMinMaxSearchStrategy<T> : IMinMaxSearchStrategy<T> where T : struct, IComparable
+    {
+        private SegmentedTree<T> segmentedTree;
+
+        public SegmentedTreeMinMaxSearchStrategy()
+        {
+            segmentedTree = new SegmentedTree<T>();
+        }
+
+        public SegmentedTreeMinMaxSearchStrategy(T[] data) : this()
+        {
+            SourceArray = data;
+        }
+
+        public T[] SourceArray
+        {
+            get => segmentedTree.SourceArray;
+            set => segmentedTree.SourceArray = value;
+        }
+
+        public void MinMaxRangeQuery(int l, int r, out double lowestValue, out double highestValue)
+        {
+            segmentedTree.MinMaxRangeQuery(l, r, out lowestValue, out highestValue);
+        }
+
+        public double SourceElement(int index)
+        {
+            return Convert.ToDouble(SourceArray[index]);
+        }
+
+        public void updateElement(int index, T newValue)
+        {
+            segmentedTree.updateElement(index, newValue);
+        }
+
+        public void updateRange(int from, int to, T[] newData, int fromData = 0)
+        {
+            segmentedTree.updateRange(from, to, newData, fromData);
+        }
+    }
+}
